Refactoring and abstraction of command handling, added a few examples
Abstracted commands so the index.js won't be full of if-else if-else statements. Now to add a new command you only need to create the corresponding .js file in the "commands" folder.  Also added a few examples, they should all work but i've not tested them yet, i'll make a fix commit if they're not working when i'll be able to test them.